scripts/hrmp: move HRMP helper from tmp; add README\n\n- Place open_to_asset_hub.ts under scripts/hrmp per PR feedback\n- Add scripts/hrmp/README with usage and security notes\n- Remove temporary tools/js/tmp files\n\nSigned-off-by: William Richter <richter@soramitsu.co.jp>
